map out contract requestor, attempt usage

+++ guests/vote-types/src/lib.rs
@@ -3,20 +3,20 @@ use alloy_sol_types::sol;
 sol! {
      #[derive(Debug)]
     struct Vote {
-        uint256 proposalId;
+        uint32 proposalId;
         address voter;
         bool choice;
     }
 
     #[derive(Debug)]
     struct VoteWitness {
-        uint256 proposalId;
+        uint32 proposalId;
         Vote[] votes;
     }
 
     #[derive(Debug)]
     struct VotePublicOutput {
-        uint256 proposalId;
+        uint32 proposalId;
         bytes32 commitmentsDigest;
         uint32 yes;
         uint32 no;
